完成发电机的添加功能

+++ app/src/main/java/edu/scut/generator/database/GeneratorDataBase.kt
@@ -1,11 +1,12 @@
 package edu.scut.generator.database
 
+import android.content.Context
 import androidx.room.Database
+import androidx.room.Room
 import androidx.room.RoomDatabase
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import java.util.*
+import edu.scut.generator.global.debug
+import kotlinx.coroutines.*
+import kotlin.system.measureTimeMillis
 
 @Database(entities = [Generator::class], version = 1, exportSchema = false)
 abstract class GeneratorDataBase : RoomDatabase() {
@@ -14,31 +15,60 @@ abstract class GeneratorDataBase : RoomDatabase() {
     @Synchronized
     fun addGenerator(vararg generator: Generator) {
         CoroutineScope(Dispatchers.IO).launch {
-            generatorDao.add(*generator)
+            if (this@GeneratorDataBase.isOpen) generatorDao.add(*generator)
         }
     }
 
     @Synchronized
     fun updateGenerator(generator: Generator) {
         CoroutineScope(Dispatchers.IO).launch {
-            generatorDao.update(generator)
+            if (this@GeneratorDataBase.isOpen) generatorDao.update(generator)
         }
     }
 
     @Synchronized
     fun deleteGenerator(generator: Generator) {
         CoroutineScope(Dispatchers.IO).launch {
-            generatorDao.delete(generator)
+            if (this@GeneratorDataBase.isOpen) generatorDao.delete(generator)
         }
     }
 
     @Synchronized
-    suspend fun getGenerator(uuid: String): Generator? {
+    fun deleteAll() {
+        CoroutineScope(Dispatchers.IO).launch {
+            if (this@GeneratorDataBase.isOpen) generatorDao.deleteAll()
+        }
+    }
+
+    @Synchronized
+    fun deleteGeneratorByUuid(uuid: String) {
+        CoroutineScope(Dispatchers.IO).launch {
+            if (this@GeneratorDataBase.isOpen) generatorDao.deleteByUuid(uuid)
+        }
+    }
+
+    @Synchronized
+    fun getGenerator(uuid: String): Generator? {
         return generatorDao.getGenerator(uuid)
     }
 
-    suspend fun getAllGenerator(): List<Generator> {
+    @Synchronized
+    fun getAllGenerator(): List<Generator> {
         return generatorDao.getAllGenerator()
+    }
+
+    companion object {
+        private lateinit var instance: GeneratorDataBase
+
+        fun init(context: Context) {
+            instance =
+                Room.databaseBuilder(context, GeneratorDataBase::class.java, "main.sql").build()
+            CoroutineScope(Dispatchers.IO).launch {
+                instance.getAllGenerator()
+            }
+        }
+
+        fun getInstance(): GeneratorDataBase = instance
     }
 
 }
